services/llm-services/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
requests
prometheus-client
pytest
httpx